ModelManagerListeners split from ModelManger
The field listeners from ModelManager was transfered to a new Singleton
class ModelManagerListeners, which is completely divided from
ModelManager, which is a god class. All the managing methods were
transferred and also all the calls to them were transferred to the new
class. After this refactoring (Splitting The God CLass) according to
InFusion the quality index is better 11.6, which means that the
rafactoring gave results, but ModelManager is still a god class and
needs more splitting.